# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProjectHub",
    "description": "Jesus Antona Espejo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/Persons/": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Find All Persons",
        "description": "Devuelve una lista de persons",
        "operationId": "find_all_persons_Persons__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Person"
                  },
                  "type": "array",
                  "title": "Response Find All Persons Persons  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Create Person",
        "description": "Crea un person y lo devuelve",
        "operationId": "create_person_Persons__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Person"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Persons/{id}": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "Find One Person",
        "description": "Devuelve el person con id pasado por par\u00e1metro",
        "operationId": "find_one_person_Persons__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del person a buscar",
              "title": "Id"
            },
            "description": "Id del person a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Persons"
        ],
        "summary": "Delete Person",
        "description": "Elimina el person con id pasado por par\u00e1metro",
        "operationId": "delete_person_Persons__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del person a eliminar",
              "title": "Id"
            },
            "description": "Id del person a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Persons"
        ],
        "summary": "Update Person",
        "description": "Actualiza el person con id pasado por par\u00e1metro y lo devuelve",
        "operationId": "update_person_Persons__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del person a actualizar",
              "title": "Id"
            },
            "description": "Id del person a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Person"
                  }
                ],
                "description": "Datos del person a actualizar",
                "title": "Person"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Companies/": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Find All Companies",
        "description": "Devuelve una lista de companies",
        "operationId": "find_all_companies_Companies__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  },
                  "type": "array",
                  "title": "Response Find All Companies Companies  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Companies"
        ],
        "summary": "Create Company",
        "description": "Crea un company y lo devuelve",
        "operationId": "create_company_Companies__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Companies/{id}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Find One Company",
        "description": "Devuelve el company con id pasado por par\u00e1metro",
        "operationId": "find_one_company_Companies__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del company a buscar",
              "title": "Id"
            },
            "description": "Id del company a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Companies"
        ],
        "summary": "Delete Company",
        "description": "Elimina el company con id pasado por par\u00e1metro",
        "operationId": "delete_company_Companies__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del company a eliminar",
              "title": "Id"
            },
            "description": "Id del company a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Companies"
        ],
        "summary": "Update Company",
        "description": "Actualiza el company con id pasado por par\u00e1metro y lo devuelve",
        "operationId": "update_company_Companies__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del company a actualizar",
              "title": "Id"
            },
            "description": "Id del company a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Company"
                  }
                ],
                "description": "Datos del company a actualizar",
                "title": "Company"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/Companies/Portfolios/{id}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Find High Portfolios",
        "description": "Devuelve los portfolios sin parent_id de company con id pasado por par\u00e1metro",
        "operationId": "find_high_portfolios_Companies_Portfolios__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id del company a buscar",
              "title": "Id"
            },
            "description": "Id del company a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Portfolio-Output"
                  },
                  "title": "Response Find High Portfolios Companies Portfolios  Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Company": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "NIF": {
            "type": "string",
            "title": "Nif"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "NIF",
          "name"
        ],
        "title": "Company"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Person": {
        "properties": {
          "DNI": {
            "type": "string",
            "title": "Dni"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "DNI",
          "name",
          "email"
        ],
        "title": "Person"
      },
...
      "Portfolio-Output": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "planned_value": {
            "type": "number",
            "title": "Planned Value"
          },
          "actual_cost": {
            "type": "number",
            "title": "Actual Cost"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "finish_date": {
            "type": "string",
            "title": "Finish Date"
          },
          "start_real_date": {
            "type": "string",
            "title": "Start Real Date"
          },
          "finish_real_date": {
            "type": "string",
            "title": "Finish Real Date"
          },
          "risk": {
            "type": "number",
            "title": "Risk"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "strategic_goal": {
            "type": "string",
            "title": "Strategic Goal"
          },
          "earned_value": {
            "type": "number",
            "title": "Earned Value"
          },
          "ROI": {
            "type": "number",
            "title": "Roi"
          },
          "cost_benefit": {
            "type": "number",
            "title": "Cost Benefit"
          }
        },
        "type": "object",
        "required": [
          "code",
          "company",
          "parent_id",
          "title",
          "description",
          "planned_value",
          "actual_cost",
          "start_date",
          "finish_date",
          "start_real_date",
          "finish_real_date",
          "risk",
          "priority",
          "strategic_goal",
          "earned_value",
          "ROI",
          "cost_benefit"
        ],
        "title": "Portfolio"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ]
}